Upgrade xmlsec to 3.0.4 and remove KeycloakFipsSecurityProvider workaround Closes #43263
