+++ docs/guides/server/fips.adoc
@@ -122,7 +122,7 @@ When starting the server, you can check that the startup log contains `KC` provi
 
 [source]
 ----
-KeycloakFipsSecurityProvider created: KC(BCFIPS version 2.0102 Approved Mode, FIPS-JVM: enabled) version 1.0
+FIPS1402Provider created: KC(BCFIPS version 2.0102 Approved Mode, FIPS-JVM: enabled)
 ----
 
 === Cryptography restrictions in strict mode
